+++ js/background.js
@@ -89,6 +89,7 @@ chrome.runtime.onStartup.addListener(function() {
 		'buildLog',
 		'builds',
 		'alarms',
+		'printLabels',
 	], function(items) {
 		if (chrome.runtime.lastError) {
 			console.error('SNAFU: Sync Get Error: %s', chrome.runtime.lastError.message);
@@ -105,6 +106,7 @@ chrome.runtime.onStartup.addListener(function() {
 			if (isVarEmpty(items.closeAlerts) === true) settingsToCreate['closeAlerts'] = true;
 			if (isVarEmpty(items.buildLog) === true) settingsToCreate['buildLog'] = false;
 			if (isVarEmpty(items.builds) === true) settingsToCreate['builds'] = {};
+			if (isVarEmpty(items.printLabels) === true) settingsToCreate['printLabels'] = false;
 			if (isVarEmpty(items.canned) === true) {
 				settingsToCreate['canned'] = {
 					'callingUser': 'Calling {INC_CUST_FNAME} at {INC_CUR_PHONE}.',